pyproject.toml
[project]
name = "madr"
version = "0.1.0"
description = ""
authors = [
    {name = "Yasmin Manuela",email = "manuelayasminsc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0" 
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)"
]

[tool.poetry]
packages = [{include = "madr", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.2"
pytest-asyncio = "^1.1.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'



[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format'
run = 'fastapi dev src/madr/app.py'
test = 'pytest -s -x --cov=fast_zero -vv'
